requirements.txt
pytest
pytest-base-url
pytest-playwright
playwright
allure-pytest
openpyxl
requests
flask

